refactor(developer): move build activities
Moves build activities under commands/build/, and build-test-data
action under commands/buildTestData. Clarifies which activites are
used where.
